+++ changelog.md
@@ -1,7 +1,5 @@
 # Changelog
 
-* No more huge mushrooms other than on mycelium
-
 ## Upcoming
 * Anvils now have only 7.5% chance to wear down when used
 * Coated Copper Tools finished
@@ -27,6 +25,7 @@
   * Is no longer rendered way up in the air and down to the bottom of the world 
   * Is now 60% transparent
 * Increased flint base damage by 0.5
+* Mushrooms are now bone meal-able only when on mycelium
 * Added Melee Protection enchantment
 * Arrows can now be made from cloth in a fletching table
 * Slight increase to Unbreaking formula (x1.29, x1.67, x2.11, x2.5 -> x1.33, x1.82, x2.5, x3.33 durability at levels I to IV)